+++ chat-client/src/client/texts/paidTier.ts
@@ -1,4 +1,4 @@
-import { ChatItem, ChatItemButton, ChatItemFormItem, ChatItemType, TextBasedFormItem } from '@aws/mynah-ui'
+import { ChatItem, ChatItemButton, ChatItemType, TextBasedFormItem } from '@aws/mynah-ui'
 
 export const plansAndPricingTitle = 'Plans &amp; Pricing'
 export const paidTierLearnMoreUrl = 'https://aws.amazon.com/q/pricing/'
@@ -13,8 +13,8 @@ export const upgradeQButton: ChatItemButton = {
     // https://github.com/aws/mynah-ui/blob/main/src/components/icon/icons/q.svg
     // https://github.com/aws/mynah-ui/blob/main/src/components/icon/icons/rocket.svg
     // icon: MynahIcons.Q,
-    description: `Upgrade to ${qProName}`,
     text: `Subscribe to ${qProName}`,
+    // description: `Upgrade to ${qProName}`,
     status: 'primary',
     disabled: false,
 }
@@ -41,29 +41,8 @@ export const continueUpgradeQButton: ChatItemButton = {
     disabled: false,
 }
 
-export const freeTierLimitCard: ChatItem = {
-    type: ChatItemType.ANSWER,
-    // Note: starts with a non-breaking space to workaround https://github.com/aws/mynah-ui/issues/349
-    title: '  Monthly request limit reached',
-    messageId: 'freetier-limit',
-    status: 'warning',
-    buttons: [],
-    icon: 'warning',
-    // iconStatus: 'success',
-    header: {
-        icon: 'warning',
-        iconStatus: 'warning',
-        body: `Upgrade to ${qProName}`,
-    },
-    canBeDismissed: false,
-    fullWidth: true,
-    body: `To increase your limit, subscribe to ${qProName}. During the upgrade, you'll be asked to link your Builder ID to the AWS account that will be billed the monthly subscription fee. Learn more about [pricing &gt;](${paidTierLearnMoreUrl})`,
-}
-
 export const freeTierLimitDirective: ChatItem = {
     type: ChatItemType.DIRECTIVE,
-    // title: '...',
-    // header: { },
     messageId: 'freetier-limit-directive',
     fullWidth: true,
     contentHorizontalAlignment: 'center',
@@ -74,23 +53,26 @@ export const freeTierLimitDirective: ChatItem = {
 /** "Banner" (sticky card) shown above the chat prompt. */
 export const freeTierLimitSticky: Partial<ChatItem> = {
     messageId: 'freetier-limit-banner',
-    title: freeTierLimitCard.title,
-    body: freeTierLimitCard.body,
+    body: `To increase your limit, subscribe to ${qProName}. During the upgrade, you'll be asked to link your Builder ID to the AWS account that will be billed the monthly subscription fee. Learn more about [pricing &gt;](${paidTierLearnMoreUrl})`,
     buttons: [upgradeQButton],
+    header: {
+        icon: 'warning',
+        iconStatus: 'warning',
+        body: '### Monthly request limit reached',
+    },
     canBeDismissed: false,
-    icon: 'warning',
-    // iconStatus: 'warning',
 }
 
 export const upgradePendingSticky: Partial<ChatItem> = {
     messageId: 'upgrade-pending-banner',
-    // Note: starts with a non-breaking space to workaround https://github.com/aws/mynah-ui/issues/349
-    body: '  Waiting for subscription status...',
-    status: 'info',
-    buttons: [],
+    body: freeTierLimitSticky.body,
+    buttons: [upgradeQButton],
+    header: {
+        icon: 'progress',
+        iconStatus: undefined,
+        body: '### Waiting for subscription status...',
+    },
     canBeDismissed: true,
-    icon: 'progress',
-    // iconStatus: 'info',
 }
 
 export const upgradeSuccessSticky: Partial<ChatItem> = {
@@ -157,7 +139,7 @@ export const paidTierPromptInput: TextBasedFormItem = {
     placeholder: '111111111111',
     type: 'textinput',
     id: 'paid-tier',
-    tooltip: `Upgrade to ${qProName}`,
+    // tooltip: `Upgrade to ${qProName}`,
     value: 'true',
     icon: 'magic',
 }